requirements.txt
requests
psycopg2-binary
python-dotenv
pandas
matplotlib
